Reuse code to delete temp files and parent directory if guid
