+++ Blaise.Api.Tests.Helpers/Files/FileSystemHelper.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.IO;
-using System.Threading;
+﻿using Blaise.Api.Extensions;
 
 namespace Blaise.Api.Tests.Helpers.Files
 {
@@ -15,55 +13,7 @@ namespace Blaise.Api.Tests.Helpers.Files
 
         public void CleanUpTempFiles(string path)
         {
-            if (!Directory.Exists(path)) return;
-
-            var directoryInfo = new DirectoryInfo(path);
-            var parentDirectory = directoryInfo.Parent?.FullName;
-            
-            if (parentDirectory == null)
-            {
-                CleanUpFiles(path);
-            }
-
-            if (Guid.TryParse(Path.GetDirectoryName(parentDirectory), out _))
-            {
-                CleanUpFiles(parentDirectory);
-                return;
-            }
-
-            CleanUpFiles(path);
-        }
-
-        private void CleanUpFiles(string path)
-        {
-            try
-            {
-                Thread.Sleep(5000);
-                DeleteDirectoryAndFilesInPath(path);
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine($"Could not cleanup folders, {e.Message}, {e}");
-            }
-        }
-
-        public static void DeleteDirectoryAndFilesInPath(string path)
-        {
-            var dirInfo = new DirectoryInfo(path);
-            foreach (var dir in dirInfo.GetDirectories())
-            {
-                foreach (var file in dir.GetFiles())
-                {
-                    file.Delete();
-                }
-
-                dir.Delete(true);
-            }
-
-            foreach (var file in Directory.GetFiles(path))
-            {
-                File.Delete(file);
-            }
+            path.CleanUpTempFiles();
         }
     }
 }
