+++ Blaise.Api/Controllers/BaseController.cs
@@ -1,11 +1,11 @@
-﻿using System;
-using System.IO;
+﻿using System.IO;
 using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Web.Http;
 using System.Web.Http.Results;
 using Blaise.Api.Contracts.Interfaces;
+using Blaise.Api.Extensions;
 using Blaise.Api.Filters;
 
 namespace Blaise.Api.Controllers
@@ -42,28 +42,7 @@ namespace Blaise.Api.Controllers
             }
             finally
             {
-                CleanUpTempFiles(filePath);
-            }
-        }
-
-        private void CleanUpTempFiles(string filePath)
-        {
-            try
-            {
-                File.Delete(filePath);
-
-                var path = Path.GetDirectoryName(filePath);
-
-                if (string.IsNullOrEmpty(path))
-                {
-                    return;
-                }
-
-                Directory.Delete(path, true);
-            }
-            catch (Exception ex)
-            {
-                _loggingService.LogInfo($"There was an error cleaning up downloaded file/folder '{filePath}' - {ex.Message}");
+                Path.GetDirectoryName(filePath).CleanUpTempFiles();
             }
         }
     }
